+++ packages/algob/src/types.ts
@@ -662,6 +662,19 @@ export interface Deployer {
   ensureCompiled: (name: string, force?: boolean, scTmplParams?: SCParams) => Promise<ASCCache>
 
   /**
+   * Returns program compiled info: bytecode (compiled program) and it's hash.
+   * @param name ASC name
+   * @param scTmplParams: Smart contract template parameters (used only when compiling PyTEAL to TEAL)
+   */
+  getCompiledASC: (name: string, scTmplParams?: SCParams) => Promise<ASCCache>
+
+  /**
+   * Returns cached program (in artifacts/cache) compiled info(bytecode, hash, filename etc).
+   * @param name ASC file name
+   */
+  getDeployedASC: (name: string) => Promise<ASCCache | undefined>
+
+  /**
    * Checks if checkpoint is deleted for a particular transaction
    * if checkpoint exists and is marked as deleted,
    * throw error(except for opt-out transactions), else pass